Restructure class locations
